// job-portal-fe-admin/src/model/jobPositionDto.ts
/**
 * job-portal
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: local
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import {ReferenceDto} from './referenceDto';
import {ReferenceUserDto} from "./referenceUserDto";

export interface JobPositionDto {
  applied?: boolean;
  company?: ReferenceUserDto;
  favourite?: boolean;
  id?: string;
  jobCategories?: Array<ReferenceDto>;
  positionName?: string;
  status?: JobPositionDto.StatusEnum;
}

export namespace JobPositionDto {
  export type StatusEnum = 'ACTIVE' | 'INACTIVE';
  export const StatusEnum = {
    ACTIVE: 'ACTIVE' as StatusEnum,
    INACTIVE: 'INACTIVE' as StatusEnum
  };
}